Fixed a bunch of warnings.

+++ neo/framework/Common.cpp
@@ -100,7 +100,7 @@ bool			com_editorActive;		//  true if an editor has focus
 #ifdef _WIN32
 HWND			com_hwndMsg = NULL;
 bool			com_outputMsg = false;
-unsigned int	com_msgID = -1;
+unsigned int	com_msgID = static_cast<unsigned int>(-1);
 #endif
 
 #ifdef __DOOM_DLL__
@@ -431,7 +431,7 @@ void idCommonLocal::VPrintf( const char *fmt, va_list args ) {
 #ifdef _WIN32
 
 	if ( com_outputMsg ) {
-		if ( com_msgID == -1 ) {
+		if ( com_msgID == static_cast<unsigned int>(-1) ) {
 			com_msgID = ::RegisterWindowMessage( DMAP_MSGID );
 			if ( !FindEditor() ) {
 				com_outputMsg = false;
